filter fucked me up
turns out using a filter is really complictaed with custom arrays so I removed the custom item xml, the Item class, and the custom adapter. I want to die because that was two days of work. Right now the necessary parts work. Currently found method of making a filter and am working on implementing it. Once I get that done, I'll fix the regular buttons and it will finally be done.